move files to proper locations
