+++ src/components/SearchBar/searchBar.tsx
@@ -7,11 +7,11 @@ import { useEffect, useState } from 'react';
 import { 
     loadMonsterList,
     selectSearchMonsterList,
-} from '../../../features/SearchMonsterListSlice';
-import { loadMonster,setShowMonsterCard } from '../../../features/MonsterCardSlice';
-import { addMonster, calcEncoutnerXP } from '../../../features/encounterSlice';
+} from '../../features/SearchMonsterListSlice';
+import { loadMonster,setShowMonsterCard } from '../../features/MonsterCardSlice';
+import { addMonster, calcEncoutnerXP } from '../../features/encounterSlice';
 
-import { AppDispatch } from '../../../app/store';
+import { AppDispatch } from '../../app/store';
 
 const SearchBar = () => {
     const dispatch = useDispatch<AppDispatch>();
